requirements.txt
tk
tkinterweb
ttkbootstrap
pillow
imageio
pandas
numpy
imagej
scyjava
markdown
pyimagej
pyinstaller
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
fonttools==4.60.1
fsspec==2024.6.1
idna==3.11
Jinja2==3.1.4
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.2
seaborn==0.13.2
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
